# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "agent",
        "title": "Probe",
        "termsOfService": "https://github.com/pnkj-kmr/probe",
        "contact": {
            "name": "Pankaj Kumar",
            "url": "https://github.com/pnkj-kmr",
            "email": "pnkj3092@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/poll/{type}/{pollPeriod}": {
            "get": {
                "description": "get totel record of poll config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Poll"
                ],
                "summary": "Poll Config Total Count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type: icmp / snmp",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Poll Period - like: 60/300",
                        "name": "pollPeriod",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "record",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "deleting the poll config records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Poll"
                ],
                "summary": "Delete All Poll Config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type: icmp / snmp",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Poll Period - like: 60/300",
                        "name": "pollPeriod",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "record",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/poll/{type}/{pollPeriod}/{id}": {
            "get": {
                "description": "get poll config data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Poll"
                ],
                "summary": "Poll Config Get View",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type: icmp / snmp",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Poll Period - like: 60/300",
                        "name": "pollPeriod",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CI ID of resource",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "poll config",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete poll config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Poll"
                ],
                "summary": "Poll Config Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type: icmp / snmp",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Poll Period - like: 60/300",
                        "name": "pollPeriod",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CI ID of resource",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "record",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}